buggy build 0.1.4.18
Changes are made in HomeActivity and corresponding recycler view adapter to rectify and correct the Firebase entries of Volunteer registration for particular event.
When in developement and testing app keeps crashing as the code segment:
241                EventPosts_position = EventPosts_position.getIndexOf();
was returning a value = -1;
due to this to correct it temporarily the code segment is replaced by:
241                EventPosts_position = 0;
this hard coded value gives a temparary solution and application stops crashing.
* This build further needed to be edited as only the hardcoded value is entered in datatbase at the runtime, wich should be dynamic by default..
* The index is defined for the keys arraylist defined prior on Adapter to support kesy storage of firebase NGO_EventPost Tree.